+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.2.0
+
+- Add `ActivityScheduler` class for finding available time slots
+- Add `WorkCalendar` class for working days and holidays
+- Add `effectiveRange` method to `ClockTimeRange` class
+- Add `median`, `mode`, `averageGap` methods to `Iterable<DateTime>` extension
+- Add `merge` method to `DartDateRange` class
+- Add `inDate` method to `ClockTime` class
+- chore: replace magic number with constants
+- chore: move errors to separate classes
+
 ## 0.1.0
 
 - Initial release